Started working on the microservices that will handle friend requests, friends, profile data of every account. Create the FriendRequest service which will handle, adding a new friend request between two accounts, deleting a friend requests and viewing all the friend requests that an account received
Update the client UI a bit for the main page, added a menu on the right
which will have the User friends and a profile preview with a button
for settings page

The AccountService changes will be refactored to be inside the upcoming
services (and this service will just validate the JWT token then
forward the request to the coresponding service that can handle it)

Some details about upcoming microservices:
- FriendService will be able to send the list of friends of an account
- FriendService will be able to delete a friendship
- FriendService will be able to add a friendship
- FriendService will check before sending data to FriendRequestService,
that the users are not friends already
- FriendRequestService will be able to add a friend request
- FriendRequestService will be able to delete a friend request
- FriendRequestService will be able to send all the friend requests of
an account
- BlockedAccountsService will be able to add a user to another user's
blocked accounts list
- BlockedAccountsService will be able to delete a user to another user's
blocked accounts list
- BlockedAccountsService will be able to send all the blocked accounts
of an account
-ProfileService will be able to manage all the details about an account,
that are not necessary for authentication (status, lastOnline etc.).
This is still to be determined.